Move util method out of task file
